Contents

Contents

User guide

Public API

User guide

SeismicWaves is a Julia package for seismic wave propagation using the finite difference method. SeismicWaves provides functions to solve the forward problem and the gradient of a misfit functional with respect to model parameters. This package additionally provides some functions to solve inverse problems using the Hamiltonian Monte Carlo (HMC) method, as part of the HMCLab framework.

Public API

SeismicWaves.SeismicWaves — Module

Module for generalized wave propagation solvers.

SeismicWaves.CPMLBoundaryConditionParameters — Type

struct CPMLBoundaryConditionParameters <: SeismicWaves.InputBoundaryConditionParameters

CPML boundary conditions parameters for wave simulations.

halo::Int64: Number of CPML grid points

rcoef::Real: Target reflection coefficient

freeboundtop::Bool: Free surface boundary condition at the top

SeismicWaves.InputParametersAcoustic — Type

struct InputParametersAcoustic{N} <: SeismicWaves.InputParameters{N}

Parameters for acoustic wave simulations.

ntimesteps::Int64: Number of time steps

dt::Real: Time step

gridsize::Tuple{Vararg{var"#s281", N}} where {N, var"#s281"<:Int64}: Grid size for each dimension

gridspacing::Tuple{Vararg{var"#s280", N}} where {N, var"#s280"<:Real}: Grid spacing in each direction

boundcond::SeismicWaves.InputBoundaryConditionParameters: Kind of boundary conditions

SeismicWaves.InputParametersAcousticVariableDensity — Type

struct InputParametersAcousticVariableDensity{N} <: SeismicWaves.InputParameters{N}

Parameters for variable-density acoustic wave simulations.

ntimesteps::Int64: Number of time steps

dt::Real: Time step

gridsize::Tuple{Vararg{var"#s281", N}} where {N, var"#s281"<:Int64}: Grid size for each dimension

gridspacing::Tuple{Vararg{var"#s280", N}} where {N, var"#s280"<:Real}: Grid spacing in each direction

boundcond::SeismicWaves.InputBoundaryConditionParameters: Kind of boundary conditions

SeismicWaves.ReflectiveBoundaryConditionParameters — Type

struct ReflectiveBoundaryConditionParameters <: SeismicWaves.InputBoundaryConditionParameters

Reflective boundary conditions parameters for wave simulations.

SeismicWaves.ScalarReceivers — Type

struct ScalarReceivers{T<:Real} <: SeismicWaves.Receivers

Type representing a multi-receiver configuration for a wave propagation shot.

positions::Matrix{<:Real}: Receiver positions

seismograms::Matrix{T} where T<:Real: Array holding seismograms (as columns)

observed::Matrix{T} where T<:Real: Array holding observed seismograms (as columns)

invcov::AbstractMatrix{T} where T<:Real: Inverse of the covariance matrix

windows::Vector{Pair{Int64, Int64}}: Windows of data used for misfit calculations

SeismicWaves.ScalarReceivers — Method

ScalarReceivers(positions, nt; observed, invcov, windows)

Create a single shot wave propagation receivers configuration from receivers positions. Default constructor for Float64.

SeismicWaves.ScalarReceivers — Method

 ScalarReceivers{T}(
   positions::Matrix{<:Real},
   nt::Integer;
   observed::Union{Matrix{T}, Nothing}=nothing,
   invcov::Union{AbstractMatrix{T}, Nothing}=nothing,
   windows::Union{Vector{Pair{Int,Int}}, Nothing}=nothing
 ) where {T <: Real}

Create a single shot wave propagation receivers configuration from receivers positions.

SeismicWaves.ScalarSources — Type

struct ScalarSources{T<:Real} <: SeismicWaves.Sources

Type representing a multi-source configuration for a wave propagation shot.

positions::Matrix{<:Real}: Source positions

tf::Matrix{T} where T<:Real: Source time function

domfreq::Real: Dominant frequency

SeismicWaves.ScalarSources — Method

ScalarSources(positions, tf, domfreq)

Create a single shot wave propagation source configuration from source positions, time-functions and a dominant frequency. Default type constructor for Float64.

SeismicWaves.ScalarSources — Method

ScalarSources{T}(positions::Matrix{<:Real}, tf::Matrix{T}, domfreq::T) where {T <: Real}

Create a single shot wave propagation source configuration from source positions, time-functions and a dominant frequency.

SeismicWaves.Shot — Type

struct Shot

Type representing a source-receiver pair, i.e., a "shot".

srcs::SeismicWaves.Sources: Structure containing the appropriate Sources for a given simulation.

recs::SeismicWaves.Receivers: Structure containing the appropriate Receivers for a given simulation.

SeismicWaves.VpAcousticCDMaterialProperty — Type

struct VpAcousticCDMaterialProperty{N} <: SeismicWaves.MaterialProperties{N}

Material properties for acoustic constant-density simulation.

vp::Array{<:Float64, N} where N: P-wave velocity

SeismicWaves.VpRhoAcousticVDMaterialProperty — Type

mutable struct VpRhoAcousticVDMaterialProperty{N} <: SeismicWaves.MaterialProperties{N}

Material properties for acoustic variable-density simulation.

vp::Array{<:Float64, N} where N: P-wave velocity

rho::Array{<:Float64, N} where N: Density

interp_method::SeismicWaves.InterpolationMethod: Interpolation method

SeismicWaves.VpRhoAcousticVDMaterialProperty — Method

VpRhoAcousticVDMaterialProperty(vp, rho; kwargs...)

Constructor to avoid specifying dimensions to create material properties for acoustic variable-density simulation.

SeismicWaves.VpRhoAcousticVDMaterialProperty — Method

VpRhoAcousticVDMaterialProperty{N}(
  vp::Array{<:Float64, N},
  rho::Array{<:Float64, N};
  interp_method::InterpolationMethod=ArithmeticAverageInterpolation(2)
) where {N}

Constructor for material properties for acoustic variable-density simulation.

SeismicWaves.WaveSimul — Type

abstract type WaveSimul{N}

Abstract type for wave simulations.

SeismicWaves.build_wavesim — Method

build_wavesim(
    params::SeismicWaves.InputParameters;
    parall,
    kwargs...
)

Builds a wave similation based on the input paramters params and keyword arguments kwargs.

Positional arguments

params::InputParameters{N}: input parameters for the simulation, where N represents the number of dimensions. They vary depending on the simulation kind (e.g., acoustic variable-density).

Keyword arguments

parall::Symbol = :threads: controls which backend is used for computation:

the CUDA.jl GPU backend performing automatic domain decomposition if set to :GPU

Base.Threads CPU threads performing automatic domain decomposition if set to :threads

Base.Threads CPU threads sending a group of sources to each thread if set to :threadpersrc

otherwise the serial version if set to :serial

gradient::Bool = false: whether the wave simulation is used for gradients computations.

check_freq::Union{<:Integer, Nothing} = nothing: if gradient = true and if specified, enables checkpointing and specifies the checkpointing frequency.

snapevery::Union{<:Integer, Nothing} = nothing: if specified, saves itermediate snapshots at the specified frequency (one every snapevery time step iteration) and return them as a vector of arrays (only for forward simulations).

infoevery::Union{<:Integer, Nothing} = nothing: if specified, logs info about the current state of simulation every infoevery time steps.

SeismicWaves.gaussderivstf — Method

gaussderivstf(t::Real, t0::Real, f0::Real) -> Real

First derivative of gaussian source time function for current time t, activation time t0 and dominating frequency f0.

SeismicWaves.gaussstf — Method

gaussstf(t::Real, t0::Real, f0::Real) -> Real

Gaussian source time function for current time t, activation time t0 and dominating frequency f0.

SeismicWaves.rickerstf — Method

rickerstf(t::Real, t0::Real, f0::Real) -> Real

Ricker source (second derivative of gaussian) source time function for current time t, activation time t0 and dominating frequency f0.

SeismicWaves.swforward! — Method

swforward!(
    params::SeismicWaves.InputParameters{N},
    matprop::SeismicWaves.MaterialProperties{N},
    shots::Vector{<:Shot};
    parall,
    snapevery,
    infoevery,
    logger
) -> Union{Nothing, Vector{AbstractArray}}

Compute forward simulation using the given input parameters params and material properties matprop on multiple shots. Receivers traces are stored in the Receivers object for each shot. See also Receivers.

Return a vector of snapshots for every shot if snapshotting is enabled.

See also Sources, Receivers.

Positional arguments

params::InputParameters{N}: input parameters for the simulation, where N represents the number of dimensions. They vary depending on the simulation kind (e.g., acoustic variable-density).

matprop::MaterialProperties{N}: material properties for the simulation, where N represents the number of dimensions. They vary depending on the simulation kind (e.g., Vp only is required for an acoustic constant-density simulation).

shots::Vector{<:Shot}: a vector whose elements are Shot structures. Each shot contains information about both source(s) and receiver(s).

Keyword arguments

parall::Symbol = :threads: controls which backend is used for computation:

the CUDA.jl GPU backend performing automatic domain decomposition if set to :GPU

Base.Threads CPU threads performing automatic domain decomposition if set to :threads

Base.Threads CPU threads sending a group of sources to each thread if set to :threadpersrc

otherwise the serial version if set to :serial

snapevery::Union{Int, Nothing} = nothing: if specified, saves itermediate snapshots at the specified frequency (one every snapevery time step iteration) and return them as a vector of arrays.

infoevery::Union{Int, Nothing} = nothing: if specified, logs info about the current state of simulation every infoevery time steps.

SeismicWaves.swforward! — Method

swforward!(
    wavesim::Union{WaveSimul{N}, Array{<:WaveSimul{N}, 1}},
    matprop::SeismicWaves.MaterialProperties{N},
    shots::Vector{<:Shot};
    logger,
    kwargs...
) -> Any

Compute forward simulation using a previously constructed WaveSimul object. Return a vector of snapshots for every shot if snapshotting is enabled.

See also Sources, Receivers.

Positional arguments

wavesim::Union{WaveSimul{N},Vector{<:WaveSimul{N}}}: input WaveSimul object containing all required information to run the simulation.

Keyword arguments

parall::Symbol = :threads: controls which backend is used for computation:

the CUDA.jl GPU backend performing automatic domain decomposition if set to :GPU

Base.Threads CPU threads performing automatic domain decomposition if set to :threads

Base.Threads CPU threads sending a group of sources to each thread if set to :threadpersrc

otherwise the serial version if set to :serial

snapevery::Union{Int, Nothing} = nothing: if specified, saves itermediate snapshots at the specified frequency (one every snapevery time step iteration) and return them as a vector of arrays.

infoevery::Union{Int, Nothing} = nothing: if specified, logs info about the current state of simulation every infoevery time steps.

SeismicWaves.swgradient! — Method

swgradient!(
    params::SeismicWaves.InputParameters{N},
    matprop::SeismicWaves.MaterialProperties{N},
    shots::Vector{<:Shot};
    parall,
    check_freq,
    infoevery,
    compute_misfit,
    misfit,
    smooth_radius,
    logger
) -> Union{Tuple{AbstractArray, Real}, AbstractArray}

Compute gradients w.r.t. model parameters using the given input parameters params and material parameters matprop on multiple shots.

The check_freq parameter controls the checkpoiting frequency for adjoint computation. If nothing, no checkpointing is performed. If greater than 2, a checkpoint is saved every check_freq time step. The optimal tradeoff value is check_freq = sqrt(nt) where nt is the number of time steps of the forward simulation. Bigger values speed up computation at the cost of using more memory.

See also Sources, Receivers, swforward!, swmisfit!.

Positional arguments

params::InputParameters{N}: input parameters for the simulation, where N represents the number of dimensions. They vary depending on the simulation kind (e.g., acoustic variable-density).

matprop::MaterialProperties{N}: material properties for the simulation, where N represents the number of dimensions. They vary depending on the simulation kind (e.g., Vp only is required for an acoustic constant-density simulation).

shots::Vector{<:Shot}: a vector whose elements are Shot structures. Each shot contains information about both source(s) and receiver(s).

Keyword arguments

parall::Symbol = :threads: controls which backend is used for computation:

the CUDA.jl GPU backend performing automatic domain decomposition if set to :GPU

Base.Threads CPU threads performing automatic domain decomposition if set to :threads

Base.Threads CPU threads sending a group of sources to each thread if set to :threadpersrc

otherwise the serial version if set to :serial

check_freq::Union{Int, Nothing} = nothing: if specified, enables checkpointing and specifies the checkpointing frequency.

infoevery::Union{Int, Nothing} = nothing: if specified, logs info about the current state of simulation every infoevery time steps.

compute_misfit::Bool = false: if true, also computes and return misfit value.

smooth_radius::Integer = 5: grid points inside a ball with radius specified by the parameter (in grid points) will have their gradient smoothed by a factor inversely proportional to their distance from sources positions.

logger::Union{Nothing,AbstractLogger}: specifies the logger to be used.

SeismicWaves.swgradient! — Method

swgradient!(
    wavesim::Union{WaveSimul{N}, Array{<:WaveSimul{N}, 1}},
    matprop::SeismicWaves.MaterialProperties{N},
    shots::Vector{<:Shot};
    logger,
    kwargs...
) -> Any

Compute gradients w.r.t. model parameters using the previously built WaveSimul. This avoids re-initializing and re-allocating several arrays in case of multiple gradient calculations.

The check_freq parameter controls the checkpoiting frequency for adjoint computation. If nothing, no checkpointing is performed. If greater than 2, a checkpoint is saved every check_freq time step. The optimal tradeoff value is check_freq = sqrt(nt) where nt is the number of time steps of the forward simulation. Bigger values speed up computation at the cost of using more memory.

See also Sources, Receivers, swforward!, swmisfit!.

Positional arguments

wavesim::Union{WaveSimul{N},Vector{<:WaveSimul{N}}}: input WaveSimul object containing all required information to run the simulation.

Keyword arguments

parall::Symbol = :threads: controls which backend is used for computation:

the CUDA.jl GPU backend performing automatic domain decomposition if set to :GPU

Base.Threads CPU threads performing automatic domain decomposition if set to :threads

Base.Threads CPU threads sending a group of sources to each thread if set to :threadpersrc

otherwise the serial version if set to :serial

check_freq::Union{Int, Nothing} = nothing: if specified, enables checkpointing and specifies the checkpointing frequency.

infoevery::Union{Int, Nothing} = nothing: if specified, logs info about the current state of simulation every infoevery time steps.

compute_misfit::Bool = false: if true, also computes and return misfit value.

smooth_radius::Integer = 5: grid points inside a ball with radius specified by the parameter (in grid points) will have their gradient smoothed by a factor inversely proportional to their distance from sources positions.

logger::Union{Nothing,AbstractLogger}: specifies the logger to be used.

SeismicWaves.swmisfit! — Method

swmisfit!(
    params::SeismicWaves.InputParameters{N},
    matprop::SeismicWaves.MaterialProperties{N},
    shots::Vector{<:Shot};
    parall,
    misfit,
    logger
) -> Real

Return the misfit w.r.t. observed data by running a forward simulation using the given input parameters params and material properties matprop on multiple shots.

Positional arguments

params::InputParameters{N}: input parameters for the simulation, where N represents the number of dimensions. They vary depending on the simulation kind (e.g., acoustic variable-density).

matprop::MaterialProperties{N}: material properties for the simulation, where N represents the number of dimensions. They vary depending on the simulation kind (e.g., Vp only is required for an acoustic constant-density simulation).

shots::Vector{<:Shot}: a vector whose elements are Shot structures. Each shot contains information about both source(s) and receiver(s).

Keyword arguments

parall::Symbol = :threads: controls which backend is used for computation:

the CUDA.jl GPU backend performing automatic domain decomposition if set to :GPU

Base.Threads CPU threads performing automatic domain decomposition if set to :threads

Base.Threads CPU threads sending a group of sources to each thread if set to :threadpersrc

otherwise the serial version if set to :serial

Receivers traces are stored in the Receivers object for each shot.

See also Sources, Receivers, swforward!.

SeismicWaves.swmisfit! — Method

swmisfit!(
    wavesim::Union{WaveSimul{N}, Array{<:WaveSimul{N}, 1}},
    matprop::SeismicWaves.MaterialProperties{N},
    shots::Vector{<:Shot};
    logger,
    kwargs...
) -> Any

Return the misfit w.r.t. observed data by running a forward simulation using the given WaveSimul object as an input.

Positional arguments

wavesim::Union{WaveSimul{N},Vector{<:WaveSimul{N}}}: input WaveSimul object containing all required information to run the simulation.

Keyword arguments

parall::Symbol = :threads: controls which backend is used for computation:

the CUDA.jl GPU backend performing automatic domain decomposition if set to :GPU

Base.Threads CPU threads performing automatic domain decomposition if set to :threads

Base.Threads CPU threads sending a group of sources to each thread if set to :threadpersrc

otherwise the serial version if set to :serial

Receivers traces are stored in the Receivers object for each shot.

See also Sources, Receivers, swforward!.